requirements.txt
colorama
yt-dlp
openai-whisper
openai
typer
pandas
openpyxl
tqdm
pyyaml
